Aligning the common block translations.
The translations of /vers/, in both C translations, are brought together.
The Fortran version uses two equivalenced variants solely to treat vedit as sometimes an integer or character.
It could as well have just used type conversion (e.g. "CHAR VEDIT" with VEDIT declared implicitly as an integer); and it will probably be modified to do so.